+++ tests/utils.py
@@ -7,9 +7,7 @@ import enum
 import os
 import pathlib
 import platform
-import random
 import re
-import rpyc
 import struct
 import subprocess
 import tempfile
@@ -32,10 +30,6 @@ DEFAULT_TARGET = TMPDIR / "default.out"
 GEF_DEFAULT_PROMPT = "gef➤  "
 GEF_DEFAULT_TEMPDIR = "/tmp/gef"
 GEF_PATH = pathlib.Path(os.getenv("GEF_PATH", "gef.py")).absolute()
-RPYC_GEF_PATH = GEF_PATH.parent / "scripts/remote_debug.py"
-RPYC_HOST = "localhost"
-RPYC_PORT = 18812
-RPYC_SPAWN_TIME = 1.0
 STRIP_ANSI_DEFAULT = True
 GDBSERVER_DEFAULT_HOST = "localhost"
 GDBSERVER_DEFAULT_PORT = 1234
@@ -71,69 +65,6 @@ class Color(enum.Enum):
 class GdbAssertionError(AssertionError):
     pass
 
-
-class RemoteGefUnitTestGeneric(unittest.TestCase):
-    """ """
-
-    def setUp(self) -> None:
-        self._coverage_file = None
-        if not hasattr(self, "_target"):
-            setattr(self, "_target", debug_target("default"))
-        else:
-            assert isinstance(self._target, pathlib.Path)  # type: ignore pylint: disable=E1101
-            assert self._target.exists()  # type: ignore pylint: disable=E1101
-        self._port = random.randint(1025, 65535)
-        self._commands = f"""
-source {GEF_PATH}
-gef config gef.debug True
-gef config gef.propagate_debug_exception True
-gef config gef.disable_color True
-source {RPYC_GEF_PATH}
-"""
-        if COVERAGE_DIR:
-            self._coverage_file = pathlib.Path(COVERAGE_DIR) / os.getenv(
-                "PYTEST_XDIST_WORKER", "gw0"
-            )
-            self._commands += f"""
-pi import coverage
-pi cov = coverage.Coverage(data_file="{self._coverage_file}", auto_data=True, branch=True)
-pi cov.start()
-"""
-
-        self._commands += f"""
-pi start_rpyc_service({self._port})
-"""
-
-        self._initfile = tempfile.NamedTemporaryFile(mode="w", delete=False)
-        self._initfile.write(self._commands)
-        self._initfile.flush()
-        self._command = [
-            "gdb",
-            "-q",
-            "-nx",
-            "-ex",
-            f"source {self._initfile.name}",
-            "--",
-            str(self._target.absolute()),  # type: ignore pylint: disable=E1101
-        ]
-        self._process = subprocess.Popen(self._command)
-        assert self._process.pid > 0
-        time.sleep(RPYC_SPAWN_TIME)
-        self._conn = rpyc.connect(
-            RPYC_HOST,
-            self._port,
-        )
-        self._gdb = self._conn.root.gdb
-        self._gef = self._conn.root.gef
-        return super().setUp()
-
-    def tearDown(self) -> None:
-        if COVERAGE_DIR:
-            self._gdb.execute("pi cov.stop()")
-            self._gdb.execute("pi cov.save()")
-        self._conn.close()
-        self._process.terminate()
-        return super().tearDown()
 
 
 class GefUnitTestGeneric(unittest.TestCase):
